prepare for saas

+++ Database/Seeders/AdminUserTableSeeder.php
@@ -5,9 +5,9 @@ namespace Modules\Core\Database\Seeders;
 use Illuminate\Database\Seeder;
 use Modules\Core\Database\Seeders\Traits\DisableForeignKeys;
 use Modules\Core\Database\Seeders\Traits\TruncateTable;
+use Modules\Core\Entities\Permission;
+use Modules\Core\Entities\Role;
 use Modules\Core\Entities\User;
-use Spatie\Permission\Models\Permission;
-use Spatie\Permission\Models\Role;
 
 class AdminUserTableSeeder extends Seeder
 {
@@ -21,21 +21,22 @@ class AdminUserTableSeeder extends Seeder
     public function run()
     {
         $this->enableForeignKeys();
-        $this->truncate('users');
-        $user = User::create([
-            'name'     => 'Admin',
-            'email'    => 'admin@gmail.com',
-            'password' => bcrypt('123456')
-        ]);
 
         $role = Role::create(['name' => 'Admin']);
         Role::create(['name' => 'User']);
-
         $permissions = Permission::pluck('id', 'id')->all();
-
         $role->syncPermissions($permissions);
 
-        $user->assignRole([$role->id]);
+        if ( ! config('core.saas_enable')) {
+            $this->truncate('users');
+            $user = User::create([
+                'name'     => 'Admin',
+                'email'    => 'admin@gmail.com',
+                'password' => bcrypt('123456')
+            ]);
+            $user->assignRole([$role->id]);
+        }
+
         $this->enableForeignKeys();
     }
 }
